+++ tests/test_files/error_files/py_file_error_empty_doctest_block.py
@@ -1,0 +1,10 @@
+"""This is an example python file"""
+
+
+class ExampleClass:
+    def method(self, *args, **kwargs):
+        """First doc str
+
+        >>>
+
+        """